LPS-71312 Move classes into correct package
